pyproject.toml
[project]
name = "dbspelunker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[evals,google]>=1.0.1",
    "pydantic-graph>=1.0.1",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.12",
]
